+++ kutana/backends/vkontakte/backend.py
@@ -27,7 +27,7 @@ class Vkontakte(Backend):
         token,
         session=None,
         requests_per_second=19,
-        api_version="5.122",
+        api_version="5.131",
         api_url="https://api.vk.com",
         **kwargs,
     ):
@@ -53,10 +53,10 @@ class Vkontakte(Backend):
 
         self.default_updates_settings = dict(
             message_new=1, message_reply=0, message_allow=0,
-            message_deny=0, message_edit=0, photo_new=0, audio_new=0,
-            video_new=0, wall_reply_new=0, wall_reply_edit=0,
-            wall_reply_delete=0, wall_reply_restore=0, wall_post_new=0,
-            wall_repost=0, board_post_new=0, board_post_edit=0,
+            message_deny=0, message_edit=0, message_event=1,
+            photo_new=0, audio_new=0, video_new=0, wall_reply_new=0,
+            wall_reply_edit=0, wall_reply_delete=0, wall_reply_restore=0,
+            wall_post_new=0, wall_repost=0, board_post_new=0, board_post_edit=0,
             board_post_restore=0, board_post_delete=0, photo_comment_new=0,
             photo_comment_edit=0, photo_comment_delete=0,
             photo_comment_restore=0, video_comment_new=0,